requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-tinymce==3.4.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
sqlparse==0.4.2
